tests/requirements.txt
wheel
cython
Pillow
pandas
scipy
parameterized
opencv-python
